backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
Pillow==10.1.0
watchdog==3.0.0
mercadopago==2.2.0
requests==2.31.0
